+++ tests/test_visitors/test_ast/test_annotations/test_return_annotations.py
@@ -5,7 +5,9 @@ import pytest
 from wemake_python_styleguide.visitors.ast.annotations import (
     LiteralNoneViolation,
     MultilineFunctionAnnotationViolation,
+    NestedAnnotationsViolation,
     WrongAnnotationVisitor,
+    WrongNestedAnnotationVisitor,
 )
 
 # Correct:
@@ -36,7 +38,47 @@ correct_embedded_return_none_annotation = """
 def function(arg) -> Union[None, Optional[int]]: ...
 """
 
+correct_unnested_literal_return = """
+def function() -> Literal[1, 2, 3]: ...
+"""
+
+correct_unnested_union_return = """
+def function() -> Union[int, str, float]: ...
+"""
+
+correct_unnested_annotated_return = """
+def function() -> Annotated[int, str, float]: ...
+"""
+
+correct_unnested_prefixed_literal_return = """
+def function() -> typing.Literal[1, 2, 3]: ...
+"""
+
 # Wrong:
+
+wrong_nested_literal_return = """
+def function() -> Literal[Literal[1, 2], 3]: ...
+"""
+
+wrong_nested_union_return = """
+def function() -> Union[Union[int, str], float]: ...
+"""
+
+wrong_nested_annotated_return = """
+def function() -> Annotated[Annotated[int, str], float]: ...
+"""
+
+wrong_deep_nested_annotated_return = """
+def function() -> Annotated[Annotated[Annotated[int]]]: ...
+"""
+
+wrong_deep_nested_literal_return = """
+def function() -> Literal[Literal[Literal[1]]]: ...
+"""
+
+wrong_deep_nested_union_return = """
+def function() -> Union[Union[Union[int]]]: ...
+"""
 
 wrong_embedded_return_none_annotation = """
 def function(arg) -> Union[Literal[None], Optional[int]]: ...
@@ -73,6 +115,14 @@ def function(
         str,
     ]
 ]: ...
+"""
+
+wrong_nested_prefixed_literal_return = """
+def function(arg: typing.Literal[typing.Literal[1]]): ...
+"""
+
+wrong_deep_nested_prefixed_literal_return = """
+def function(arg: typing.Literal[typing.Literal[typing.Literal[1]]]): ...
 """
 
 
@@ -118,6 +168,53 @@ def test_wrong_literal_none_return_annotation(
 
 
 @pytest.mark.parametrize('code', [
+    wrong_nested_literal_return,
+    wrong_nested_union_return,
+    wrong_nested_prefixed_literal_return,
+    wrong_nested_annotated_return,
+])
+def test_wrong_nested_return_annotations(
+    assert_errors,
+    parse_ast_tree,
+    code,
+    default_options,
+    mode,
+):
+    """Ensures that nested return ``Literal`` and ``Union`` is forbidden."""
+    tree = parse_ast_tree(mode(code))
+
+    visitor = WrongNestedAnnotationVisitor(default_options, tree=tree)
+    visitor.run()
+
+    assert_errors(visitor, [NestedAnnotationsViolation])
+
+
+@pytest.mark.parametrize('code', [
+    wrong_deep_nested_literal_return,
+    wrong_deep_nested_union_return,
+    wrong_deep_nested_annotated_return,
+    wrong_deep_nested_prefixed_literal_return,
+])
+def test_wrong_deep_nested_return_annotations(
+    assert_errors,
+    parse_ast_tree,
+    code,
+    default_options,
+    mode,
+):
+    """Ensures that nested return ``Literal`` and ``Union`` is forbidden."""
+    tree = parse_ast_tree(mode(code))
+
+    visitor = WrongNestedAnnotationVisitor(default_options, tree=tree)
+    visitor.run()
+
+    assert_errors(visitor, [
+        NestedAnnotationsViolation,
+        NestedAnnotationsViolation,
+    ])
+
+
+@pytest.mark.parametrize('code', [
     correct_function_without_annotations,
     correct_simple_return,
     correct_compound_return,
@@ -136,6 +233,28 @@ def test_correct_return_annotation(
     tree = parse_ast_tree(mode(code))
 
     visitor = WrongAnnotationVisitor(default_options, tree=tree)
+    visitor.run()
+
+    assert_errors(visitor, [])
+
+
+@pytest.mark.parametrize('code', [
+    correct_unnested_literal_return,
+    correct_unnested_union_return,
+    correct_unnested_annotated_return,
+    correct_unnested_prefixed_literal_return,
+])
+def test_correct_unnested_annotation(
+    assert_errors,
+    parse_ast_tree,
+    code,
+    default_options,
+    mode,
+):
+    """Ensures that it is possible to use correct unnested annotations."""
+    tree = parse_ast_tree(mode(code))
+
+    visitor = WrongNestedAnnotationVisitor(default_options, tree=tree)
     visitor.run()
 
     assert_errors(visitor, [])